+++ hypervisor/boot/include/reloc.h
@@ -6,12 +6,29 @@
 #ifndef RELOCATE_H
 #define RELOCATE_H
 
-extern void relocate(void);
-extern uint64_t get_hv_image_delta(void);
-extern uint64_t get_hv_image_base(void);
-extern uint64_t get_hv_image_size(void);
+#include <elf.h>
+#include <asm/boot/ld_sym.h>
 
-/* external symbols that are helpful for relocation */
-extern uint8_t		_DYNAMIC[1];
+uint64_t arch_get_hv_image_delta(void);
+
+static inline uint64_t get_hv_image_delta(void)
+{
+	return arch_get_hv_image_delta();
+}
+
+/* get the actual Hypervisor load address (HVA) */
+static inline uint64_t get_hv_image_base(void)
+{
+	return (get_hv_image_delta() + CONFIG_HV_RAM_START);
+}
+
+static inline uint64_t get_hv_image_size(void)
+{
+	return (uint64_t)(&ld_ram_end - &ld_ram_start);
+}
+
+#ifdef CONFIG_RELOC
+extern void relocate(struct Elf64_Dyn *dynamic);
+#endif
 
 #endif /* RELOCATE_H */
